Add unit tests for remove<T> in common/helpers
